+++ src/WeightedEdge.java
@@ -3,16 +3,15 @@ package src;
 public class WeightedEdge {
 	int v1;
 	int v2;
-	int weight;
-	public WeightedEdge(int v1, int v2, int weight) {
+	double weight;
+	public WeightedEdge(int v1, int v2, double weight) {
 		super();
 		this.v1 = v1;
 		this.v2 = v2;
 		this.weight = weight;
 	}
-	public void addWeight(int w) {
-		//weight+=weight;
-		weight+=w; // MELI
+	public void addWeight(double w) {
+		weight+=w;
 	}
 	public int getV1() {
 		return v1;
@@ -20,8 +19,7 @@ public class WeightedEdge {
 	public int getV2() {
 		return v2;
 	}
-	public int getWeight() {
+	public double getWeight() {
 		return weight;
-	}
-	
+	}	
 }
